Move TreeGrid specific demos to its own class
That way the TreeGrid demos can be isolated in vaadin.com.
